requirements.txt
wheel
requests
cherrypy
pillow